refactor chapter 3 code
