+++ c2_NetworkDoubleSpends/2.3-Latency-Based_Consensus/timestampTest.js
@@ -39,7 +39,7 @@ genesis[wallets[0].address] = {
 const nodes = []
 // Create new nodes based on our wallets, and connect them to the network
 for (let i = 0; i < numNodes; i++) {
-  nodes.push(new FaultTolerant(wallets[i], JSON.parse(JSON.stringify(genesis)), network, delta = 7))
+  nodes.push(new FaultTolerant(wallets[i], JSON.parse(JSON.stringify(genesis)), network, delta = 5))
   // Connect everyone to everyone
   network.connectPeer(nodes[i], numConnections = 3)
 }
