+++ c3_ProofOfWork/nodeAgent.js
@@ -1,0 +1,101 @@
+var EthCrypto = require('eth-crypto')
+
+function getTxHash (tx) {
+  return EthCrypto.hash.keccak256(JSON.stringify(tx))
+}
+
+class Node {
+  constructor (wallet, genesis, network) {
+    // Blockchain identity
+    this.wallet = wallet
+    // P2P Node identity -- used for connecting to peers
+    this.p2pNodeId = EthCrypto.createIdentity()
+    this.pid = this.p2pNodeId.address
+    this.network = network
+    this.state = genesis
+    this.transactions = []
+    this.invalidNonceTxs = {}
+  }
+
+  onReceive (tx) {
+    if (this.transactions.includes(tx)) {
+      return
+    }
+    this.transactions.push(tx)
+    this.applyTransaction(tx)
+    this.network.broadcast(this.pid, tx)
+    this.applyInvalidNonceTxs(tx.contents.from)
+  }
+
+  applyInvalidNonceTxs (address) {
+    const targetNonce = this.state[address].nonce
+    if (address in this.invalidNonceTxs && targetNonce in this.invalidNonceTxs[address]) {
+      this.applyTransaction(this.invalidNonceTxs[address][targetNonce])
+      delete this.invalidNonceTxs[address][targetNonce]
+      this.applyInvalidNonceTxs(address)
+    }
+  }
+
+  tick () {}
+
+  generateTx (to, amount) {
+    const unsignedTx = {
+      type: 'send',
+      amount: amount,
+      from: this.wallet.address,
+      to: to,
+      nonce: this.state[this.wallet.address].nonce
+    }
+    const tx = {
+      contents: unsignedTx,
+      sig: EthCrypto.sign(this.wallet.privateKey, getTxHash(unsignedTx))
+    }
+    return tx
+  }
+
+  applyTransaction (tx) {
+    // Check the from address matches the signature
+    const signer = EthCrypto.recover(tx.sig, getTxHash(tx.contents))
+    if (signer !== tx.contents.from) {
+      throw new Error('Invalid signature!')
+    }
+    // If we don't have a record for this address, create one
+    if (!(tx.contents.to in this.state)) {
+      this.state[[tx.contents.to]] = {
+        balance: 0,
+        nonce: 0
+      }
+    }
+    // Check that the nonce is correct for replay protection
+    if (tx.contents.nonce !== this.state[[tx.contents.from]].nonce) {
+      // If it isn't correct, then we should add the transaction to invalidNonceTxs
+      if (!(tx.contents.from in this.invalidNonceTxs)) {
+        this.invalidNonceTxs[tx.contents.from] = {}
+      }
+      this.invalidNonceTxs[tx.contents.from][tx.contents.nonce] = tx
+      return
+    }
+    if (tx.contents.type === 'send') { // Send coins
+      if (this.state[[tx.contents.from]].balance - tx.contents.amount < 0) {
+        console.log('not enough money...yet')
+        return
+        // throw new Error('Not enough money!')
+      }
+      this.state[[tx.contents.from]].balance -= tx.contents.amount
+      this.state[[tx.contents.to]].balance += tx.contents.amount
+    } else if (tx.contents.type === 'mint') { //Block reward
+      // Check the from address matches the signature
+      const signer = EthCrypto.recover(tx.sig, getTxHash(tx.contents))
+      if (signer !== tx.contents.to) {
+        throw new Error('Invalid signature in mint tx!')
+      }
+      // Update state
+      this.state[[tx.contents.to]].balance += tx.contents.amount
+    } else {
+      throw new Error('Invalid transaction type!')
+    }
+    this.state[[tx.contents.from]].nonce += 1
+  }
+}
+
+module.exports = {Node, getTxHash}
